+++ app/src/main/java/com/android/go/sopt/winey/presentation/main/feed/upload/PhotoFragment.kt
@@ -2,12 +2,16 @@ package com.android.go.sopt.winey.presentation.main.feed.upload
 
 import android.os.Bundle
 import android.view.View
+import androidx.activity.OnBackPressedCallback
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.commit
 import com.android.go.sopt.winey.R
 import com.android.go.sopt.winey.databinding.FragmentPhotoBinding
 import com.android.go.sopt.winey.util.binding.BindingFragment
+import com.android.go.sopt.winey.util.context.stringOf
+import com.android.go.sopt.winey.util.fragment.WineyDialogFragment
+import com.android.go.sopt.winey.util.fragment.stringOf
 
 class PhotoFragment : BindingFragment<FragmentPhotoBinding>(R.layout.fragment_photo) {
     private val uploadViewModel by activityViewModels<UploadViewModel>()
@@ -19,6 +23,28 @@ class PhotoFragment : BindingFragment<FragmentPhotoBinding>(R.layout.fragment_ph
         initImageSelectButtonClickListener()
         initNextButtonClickListener()
         initCloseButtonClickListener()
+        registerBackPressedCallback()
+    }
+
+    private fun registerBackPressedCallback() {
+        val callback = object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                showAlertDialog()
+            }
+        }
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, callback)
+    }
+
+    private fun showAlertDialog() {
+        val dialog = WineyDialogFragment(
+            stringOf(R.string.upload_dialog_title),
+            stringOf(R.string.upload_dialog_subtitle),
+            stringOf(R.string.upload_dialog_negative_button),
+            stringOf(R.string.upload_dialog_positive_button),
+            handleNegativeButton = {},
+            handlePositiveButton = { requireActivity().finish() }
+        )
+        dialog.show(parentFragmentManager, TAG_UPLOAD_DIALOG)
     }
 
     private fun initImageSelectButtonClickListener() {
@@ -64,7 +90,7 @@ class PhotoFragment : BindingFragment<FragmentPhotoBinding>(R.layout.fragment_ph
     }
 
     companion object {
-        private const val ARGS_PHOTO_KEY = "photo"
         private const val IMAGE_FILE = "image/*"
+        private const val TAG_UPLOAD_DIALOG = "UPLOAD_DIALOG"
     }
 }
